pyproject.toml
[tool.poetry]
name = "hse-aiahp-baseline"
version = "0.1.0"
description = "Baseline solution of HSE AI Assistant Hack: Python [https://www.hse.ru/ai-assistant-hack-python/]"
authors = ["Maxim Kirilyuk <werserk@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
torch = { version = "2.5.0", source="torch"}
transformers = "^4.45.2"
openpyxl = "^3.1.5"
llama-cpp-python = "^0.3"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu124"
secondary = true

[tool.poetry.group.dev.dependencies]
black = { extras = ["jupyter"], version = "^24.10.0" }
pre-commit = "^4.0.1"
jupyter = "^1.1.1"
tqdm = "^4.66.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120